+++ registries.go
@@ -1,0 +1,8 @@
+package oci
+
+// Default OCI registry base paths for each Klaus artifact type.
+const (
+	DefaultPluginRegistry      = "gsoci.azurecr.io/giantswarm/klaus-plugins"
+	DefaultPersonalityRegistry = "gsoci.azurecr.io/giantswarm/klaus-personalities"
+	DefaultToolchainRegistry   = "gsoci.azurecr.io/giantswarm"
+)